+++ go/vt/vtgate/shard_conn.go
@@ -188,7 +188,7 @@ func (sdc *ShardConn) Rollback(ctx context.Context, transactionID int64) (err er
 }
 
 // SplitQuery splits a query into sub queries. The retry rules are the same as Execute.
-func (sdc *ShardConn) SplitQuery(ctx context.Context, sql string, bindVariables map[string]interface{}, splitColumn string, splitCount int) (queries []tproto.QuerySplit, err error) {
+func (sdc *ShardConn) SplitQuery(ctx context.Context, sql string, bindVariables map[string]interface{}, splitColumn string, splitCount int64) (queries []tproto.QuerySplit, err error) {
 	err = sdc.withRetry(ctx, func(conn tabletconn.TabletConn) error {
 		var innerErr error
 		queries, innerErr = conn.SplitQuery(ctx, tproto.BoundQuery{
