+++ go/vt/vtgate/vtgateservice/interface.go
@@ -39,7 +39,7 @@ type VTGateService interface {
 	Rollback(ctx context.Context, session *vtgatepb.Session) error
 
 	// Map Reduce support
-	SplitQuery(ctx context.Context, keyspace string, sql string, bindVariables map[string]interface{}, splitColumn string, splitCount int) ([]*vtgatepb.SplitQueryResponse_Part, error)
+	SplitQuery(ctx context.Context, keyspace string, sql string, bindVariables map[string]interface{}, splitColumn string, splitCount int64) ([]*vtgatepb.SplitQueryResponse_Part, error)
 
 	// Topology support
 	GetSrvKeyspace(ctx context.Context, keyspace string) (*topodatapb.SrvKeyspace, error)
